Thu, Mar 16, 2023, 11:27 AM +00:00

+++ calculator.c
@@ -1,7 +1,7 @@
 #include <cs50.h>
 #include <stdio.h>
 
-long main(void)
+int main(void)
 {
    long x = get_long("x: ");
    long y = get_long("y: ");
